Fixed #83
After a long investigation I was able to find out what caused this issue and how to fix it (simple fix)
For some reason, the menu buttons view controller immediately activating does not give enough time for the ModalView on the left screen (any ModalView even if it isnt in gameplay setup) to finish. Giving it an animation fixes it.