Fix: FloatingScreen handle does not work in level
A new instance VRPointer now gets initialized again in Game Scene (and destroyed after the level is over). This renders the single reference of VRPointer from the menu useless and FloatingScreen handle does not work as a result.
To fix this, I harmony patch the OnEnable for VRPointer and pass this reference along to the FloatingScreen to update it.