Fixed Refresh() exception for GameplaySetup
Explanation: If we go into campaign/multi/custom version of GameplaySetup first, select something from the mods tab and go back to solo, this exception would occur. Upon inspection, I found out it is because the vanilla tab is not the first tab selected when we enter Solo again so it is not active and enabled, causing issues with base game logic. I performed this cleanup on deactivate.