Added slider with buttons
Changes to SliderSetting and ListSliderSetting: The slider we used for cloning is now replaced with the controller settings slider instead (requiring less things to clone as a result too to get the text). The slider buttons are also cloned: so sliders could be precisely changed with buttons on either end. To enable it, "show-buttons" needs to be set to true.